First: Reorganize Mario's acmd
