+++ src/fighter/mario/acmd/mod.rs
@@ -1,0 +1,11 @@
+mod smash_attacks;
+mod aerials;
+mod specials;
+mod misc;
+
+pub fn install() {
+    smash_attacks::install();
+    aerials::install();
+    specials::install();
+    misc::install();
+}
